+++ src/ScriptSymlinker.php
@@ -12,7 +12,7 @@
 namespace Ajgl\Composer;
 
 use Composer\Script\Event;
-use Symfony\Component\Filesystem\Filesystem;
+use Composer\Util\Filesystem;
 
 /**
  * Script to symlink resources installed with composer.
@@ -24,15 +24,40 @@ class ScriptSymlinker
     public static function createSymlinks(Event $event)
     {
         $symlinks = static::getSymlinks($event);
-        $vendorDir = $event->getComposer()->getConfig()->get('vendor-dir');
         $fs = new Filesystem();
 
-        foreach ($symlinks as $package => $symlinksDefinition) {
-            foreach ($symlinksDefinition as $origin => $target) {
-                $originPath = realpath($vendorDir)."/$package/$origin";
-                $targetPath = realpath($vendorDir)."/$target";
-                $event->getIO()->write("Symlinking <comment>$originPath</comment> --> <comment>$targetPath</comment>");
-                $fs->symlink($originPath, $targetPath);
+        foreach ($event->getComposer()->getRepositoryManager()->getLocalRepository()->getPackages() as $package) {
+            if (isset($symlinks[$package->getName()])) {
+                $packageDir = $event->getComposer()->getInstallationManager()->getInstallPath($package);
+
+                $symlinkDefinitions = $symlinks[$package->getName()];
+                foreach ($symlinkDefinitions as $target => $link) {
+                    if ($fs->isAbsolutePath($target)) {
+                        throw new \InvalidArgumentException(
+                            "Invalid symlink target path '$target' for package'{$package->getName()}'."
+                            .' It must be relative.'
+                        );
+                    }
+                    if ($fs->isAbsolutePath($link)) {
+                        throw new \InvalidArgumentException(
+                            "Invalid symlink link path '$link' for package'{$package->getName()}'."
+                            .' It must be relative.'
+                        );
+                    }
+
+                    $targetPath = $packageDir.DIRECTORY_SEPARATOR.$target;
+                    $linkPath = getcwd().DIRECTORY_SEPARATOR.$link;
+
+                    if (!file_exists($targetPath)) {
+                        throw new \RuntimeException(
+                            "The target path '$targetPath' for package'{$package->getName()}' does not exist."
+                        );
+                    }
+
+                    $event->getIO()->write("Symlinking <comment>$targetPath</comment> to <comment>$linkPath</comment>");
+                    $fs->ensureDirectoryExists(dirname($linkPath));
+                    $fs->relativeSymlink($targetPath, $linkPath);
+                }
             }
         }
     }
@@ -40,6 +65,7 @@ class ScriptSymlinker
     protected static function getSymlinks(Event $event)
     {
         $options = $event->getComposer()->getPackage()->getExtra();
+
         $symlinks = array();
 
         if (isset($options['ajgl-symlinks']) && is_array($options['ajgl-symlinks'])) {